+++ frontend/src/admin/Panel/Post/Post.tsx
@@ -1,4 +1,5 @@
 import { Component, For, Setter } from 'solid-js';
+import { adminDeletePost } from '../../../api/adminAPI';
 
 import { IMeta } from '../../../api/siteAPI';
 
@@ -8,6 +9,12 @@ interface IProps {
 }
 
 const Post: Component<IProps> = (props) => {
+  const deletePost = async () => {
+    try {
+      await adminDeletePost(props.meta.id);
+    } catch (error) {}
+  };
+
   return (
     <div style={{ 'margin-bottom': '30px' }}>
       <p>{props.meta.id}</p>
@@ -16,6 +23,7 @@ const Post: Component<IProps> = (props) => {
       <p>{props.meta.series}</p>
       <For each={props.meta.categories}>{(c) => <span>{c} </span>}</For>
       <button onClick={() => props.setSelectedPost(props.meta.id)}>Edit</button>
+      <button onClick={deletePost}>Delete</button>
     </div>
   );
 };